Move enum utils to options selector utils
