+++ lib/streamlit/elements/widgets/multiselect.py
@@ -24,6 +24,7 @@ from streamlit.elements.lib.options_selector_utils import (
     check_and_convert_to_indices,
     convert_to_sequence_and_check_comparable,
     get_default_indices,
+    maybe_coerce_enum_sequence,
 )
 from streamlit.elements.lib.policies import (
     check_widget_policies,
@@ -34,7 +35,6 @@ from streamlit.elements.lib.utils import (
     LabelVisibility,
     compute_and_register_element_id,
     get_label_visibility_proto_value,
-    maybe_coerce_enum_sequence,
     save_for_app_testing,
     to_key,
 )
