+++ lib/streamlit/elements/widgets/radio.py
@@ -20,6 +20,7 @@ from typing import TYPE_CHECKING, Any, Callable, Generic, Sequence, cast, overlo
 
 from streamlit.dataframe_util import OptionSequence, convert_anything_to_list
 from streamlit.elements.lib.form_utils import current_form_id
+from streamlit.elements.lib.options_selector_utils import index_, maybe_coerce_enum
 from streamlit.elements.lib.policies import (
     check_widget_policies,
     maybe_raise_label_warnings,
@@ -29,7 +30,6 @@ from streamlit.elements.lib.utils import (
     LabelVisibility,
     compute_and_register_element_id,
     get_label_visibility_proto_value,
-    maybe_coerce_enum,
     save_for_app_testing,
     to_key,
 )
@@ -48,7 +48,6 @@ from streamlit.type_util import (
     T,
     check_python_comparable,
 )
-from streamlit.util import index_
 
 if TYPE_CHECKING:
     from streamlit.delta_generator import DeltaGenerator
