+++ lib/streamlit/util.py
@@ -22,13 +22,12 @@ import hashlib
 import os
 import subprocess
 import sys
-from typing import Any, Callable, Final, Iterable, Mapping, TypeVar
+from typing import Any, Callable, Final, Mapping, TypeVar
 
 from streamlit import env_util
 
 # URL of Streamlit's help page.
 HELP_DOC: Final = "https://docs.streamlit.io/"
-FLOAT_EQUALITY_EPSILON: Final[float] = 0.000000000005
 
 # Due to security issue in md5 and sha1, usedforsecurity
 # argument is added to hashlib for python versions higher than 3.8
@@ -128,31 +127,6 @@ def repr_(self: Any) -> str:
 
 
 _Value = TypeVar("_Value")
-
-
-def index_(iterable: Iterable[_Value], x: _Value) -> int:
-    """Return zero-based index of the first item whose value is equal to x.
-    Raises a ValueError if there is no such item.
-
-    We need a custom implementation instead of the built-in list .index() to
-    be compatible with NumPy array and Pandas Series.
-
-    Parameters
-    ----------
-    iterable : list, tuple, numpy.ndarray, pandas.Series
-    x : Any
-
-    Returns
-    -------
-    int
-    """
-    for i, value in enumerate(iterable):
-        if x == value:
-            return i
-        elif isinstance(value, float) and isinstance(x, float):
-            if abs(x - value) < FLOAT_EQUALITY_EPSILON:
-                return i
-    raise ValueError(f"{str(x)} is not in iterable")
 
 
 _Key = TypeVar("_Key", bound=str)
